pyproject.toml
[tool.poetry]
name = "djangoslackproject"
version = "0.1.0"
description = ""
authors = ["SORARAwo4649 <k11274649@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.8.5"
django = "^3.0.9"
slackclient = "^2.7.3"
slackweb = "^1.0.5"
requests = "^2.24.0"
certifi = "^2020.6.20"
django-bootstrap4 = "^2.2.0"
dj-database-url = "^0.5.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
